requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.2
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==1.10.8
django-annoying==0.10.4
django-bootstrap3==9.1.0
django-ckeditor==5.4.0
django-contrib-comments==1.8.0
django-extensions==2.0.5
django-guardian==1.4.9
django-hitcount==1.2.4
django-js-asset==1.0.0
django-simplepagination==0.0.1
django-userena==2.0.1
easy-thumbnails==2.5
filebrowser-safe==0.4.7
future==0.16.0
grappelli-safe==0.4.7
html2text==2014.12.29
html5lib==1.0.1
idna==2.6
Mezzanine==4.2.3
oauthlib==2.0.6
Pillow==5.0.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tzlocal==1.5.1
urllib3==1.22
webencodings==0.5.1
